+++ database/migrations/0001_01_01_000001_create_cache_table.php
@@ -8,30 +8,28 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::create('cache', function (Blueprint $table): void {
+        Schema::create('cache', static function (Blueprint $table): void {
             $table->string('key')->primary();
+
             $table->mediumText('value');
             $table->integer('expiration');
         });
 
-        Schema::create('cache_locks', function (Blueprint $table): void {
+        Schema::create('cache_locks', static function (Blueprint $table): void {
             $table->string('key')->primary();
+
             $table->string('owner');
             $table->integer('expiration');
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
-        Schema::dropIfExists('cache');
-        Schema::dropIfExists('cache_locks');
+        if (! app()->isProduction()) {
+            Schema::dropIfExists('cache');
+            Schema::dropIfExists('cache_locks');
+        }
     }
 };
